Add support for real, bv8, bv16, bv32, bv64
A fully-parameterized bit vector type doesn't work well at all stages of
the pipeline at the moment, but could potentially in the future. The
Lambda dialect, at least, has a `.bitvec n` constructor in `LMonoTy`,
but most of the rest of the pipeline treats each bit vector type
entirely independently.