Add loop removal transformation for Boogie
Based heavily on what @andrewmwells-amazon did for C_Simp.

Ultimately, C_Simp should probably just translate to Boogie, with loops,
and then call this tranformation.

In the process, I consolidated and generalized a lot of the factory
functions, because I happened to want to refer to them more robustly in
the loop transformation.