Translate Boogie arithmetic on type aliases
Translation of Boogie arithmetic operators requires that the type being
operated on is a type that supports arithmetic. In the presence of type
aliasing, this may require dealiasing during transformation, which now
occurs.

Along the way, fix some of the examples to adapt to new function
signatures, and add several previously-missing examples to the top-level
Boogie.Examples module.

Also fix translation of the modulus operation.