Add support for multiple triggers
Now, in the Boogie dialect, it's possible to write quantified formulas
like the following:

    forall x : int, y : int :: { g(x, y) } { f(x) } g(x, y) < f(x);
    forall x : int, y : int :: { g(x, y), f(x) } g(x, y) < f(x);

Each quantifier can have multiple trigger groups, and each group can
have multiple expressions. The indended semantics according to the
SMT-Lib standard is that each group is an alternative pattern that can
cause the quantifier to be instantiated. To instantiate based on a
specific group, the solver should find a ground term that matches each
expression in the group.